[FIX] Combining invoices keeps the sale.order.line reference intact
Because self.env['sale.order'] returns an empty record set, this is evaluated to False.
This in turn never calls the code to restore the link between sale order and invoice.
Because this code was never called in version 9 and above, it never encountered the issue
that the the field invoice_ids is now a computed field without a search.
Searching on that field returned all the sale.orders.
The linked sale_line_ids are now copied like the rest of the fields.