[FIX] account_invoice_triple_discount: cache can be reset
in some rare case, happening in a non deterministic way AFAICT, the cache can be reset,
resulting in wrong computations, because odoo will repopulate the cache with the data from the
database.

This patch attempts to detect this, and to fix the issue by relauching the method.

This is ugly, but in my opinion, messing with the cache as this module does is
a really bad idea.