[IMP] sale_order_invoicing_queued: Make it compatible with grouping criteria
We compatibilize this module with `sale_order_invoicing_grouping_criteria` one
checking explicitly if there's a method with the name that the other module
declares, and if so, calling it for getting the grouping criteria.

This is the only valid approach for now, as we don't have any hook for getting that
criteria.