[12.0][ADD] - substitution rules for transmit method
Sometimes a transmission method defined by default for the partner is not
applicable. For example, when the transmission method is e-mail and the
partner's e-mail is not set. This addon allows you to define substitution
rules when an invoice matches certain criteria.

This module can also be used when it is not easy to define the partner's
preference for the transmission method or the partner's preference is not
legally permitted. For example, in some countries, invoices with payment
direct debit method must be sent to the customer by post.