account_invoice_transmit_method: Remove arg with wrong syntax
In the initial dev of the module, I made a typo in the definition of the
fields customer_invoice_transmit_method_id and supplier_invoice_transmit_method_id
on res.partner: I wrote "company_dependant=True" instead of
"company_dependent=True".
As a result, these 2 fields were not company-specific.
So we have 2 options:
1) Fix the typo and write migration scripts
2) Keep these 2 fields with company_dependent=False, as nobody as
complained so far about the fact that the value of this parameter didn't
vary from one company to another.

This commit implement option 2.