[FIX] sale_order_invoicing_grouping_criteria: Don't inject grouping criteria, but slice super calls
It turns out that the grouping criteria is applied over the invoice vals dictionary
list instead on sales order so you can't group by any field in sales order,
including the trick of slicing by ID for not grouping at all.

We then restore and adapt the v12 code, that slices the recordset in pieces
according the grouping criteria with fields obtained in the sales orders.