[FIX] account_invoice_supplierinfo_update: rounding
We shouldn't roung the price unit, as we'd lose precision when writing
back to the supplierinfo from the original unit price item which is the
one we want to update. For example:

- We buy an item at 1,2345 (decimal precision of 4 digits for product
price)
- The product is correctly invoiced at 1,2345 (rounded in subtotal)
- We should update the item supplierinfo at that same price. Otherwise,
if we round to the currency digits (usally 2) we'd get something like:
1,23 in the supplierinfo and in the next purchase we'd be losing those
decimal which in large amounts of units translates into a lot of money.

TT32538