[FIX] purchase_stock_picking_return_invoicing: Limit update scope
Previous to this commit, every time there's a difference between the
expected quantities (ordered/received vs invoiced), the invoice status
was overwritten, not allowing other modules like
`purchase_force_invoiced` to act.

Now, we limit the scope of the update to only the condition that this
module is adding: when ordered/received is lower than the invoiced
quantity.

TT38533