[IMP] sale_timesheet_invoice_description: Add a unit test
In order to cover the previous fix for the invoice line sequences,
the unit test for invoice split by timesheets has to cover an order
with multiple tasks. Only this way, the correct order of section and
product lines in an invoice can be tested.
Hence, another sales order, with two products, is created, resulting
in a separate task, for which then according timesheets are created.

When copying a timesheet, its associated sale order line is getting
associated also with the copy. But the copy may have its own task, which
has its own sale order line, which should be assigned to the timesheet.
The OCA module "sale_timesheet_line_exclude" does that (by its overwritten
method `_timesheet_postprocess()`). We need to do it manually, e.g. by
simply calling `_compute_so_line()` on the timesheet copies.