[FIX] sale_timesheet_invoice_description: Split for multiple invoices
If an invoice for a sale order had been created with the split feature and
then modified, e.g. by removing split lines or changing their quantities,
another invoice can be created to invoice those quantities missing on that
first invoice. This commit tries to correctly handle the split for this
additional invoice.
This is done by taking into account the quantities of timesheets already
invoiced on the first invoice. The quantities on the new invoice have to
be reduced by them. This implies that timesheets already fully invoiced
must not be part of the new invoice.