[FIX] sale_timesheet_invoice_description: Make an internal field readonly and add a technical comment
Note that the value of this field, the split description option on model
"account.move.line", gets copied from model "sale.order". The field is defined
as `Char`, even though its source field is `Selection`. This way, code
duplication can be avoided, namely the need to specify the `selection`
parameter. Note that the field is not available to the Odoo users in the UI,
but rather used just "internally" in the code. And there, on code as well as
database level, the values of `Selection` and `Char` fields are just the same,
namely strings. Hence, this works just fine.
Just in case someone decides to make the field available to the Odoo users (by
adding it to a view), we declare it to be readonly, as its value should not be
changed.