[FIX] sale_timesheet_invoice_description: Prevent unbalanced invoice
Invoices must not and cannot be unbalanced, otherwise an exception is
raised. It is done here by not checking the balance while splitting the
account move and its lines, and then recomputing the dynamic lines
(subtotal, total and taxes) after the splitting.

Note that an additional balance check could implemented to be done
afterwards, i.e., `_check_balanced` after the `_recompute_dynamic_lines`.