[FIX] account_invoice_fiscal_position_update: fix tax incl. to tax excl.
Fixes #1896

This change causes the subtotal amount to remain the same when a fiscal position
change triggers a tax update from tax included in price to tax excluded from
price.

Tests are added for conversion between tax excluded to tax included as well, but
in this case the behaviour is not 'fixed'. The standard behaviour of Odoo in
case of tax excl. to tax incl. is to adjust the subtotal amount. The same
behaviour is present in the function that this module provides. It might not
be what users expect, but it seems that Odoo does not support to preserve the
subtotal amount in this case.