[FIX] account_invoice_purchase_picking_selection: Use commercial_partner_id instead of partner_id
When a contact has children and one is marked as the invoice address, the partner_id is, by default, set to this invoice contact. As a result, the pickings only display records related to that invoice recipient.
With this commit, we now use commercial_partner_id instead, ensuring all related records are correctly shown.