[IMP] account_invoice_supplier_self_invoice: Reference auto generation optional
Makes the generation of a client reference on self billed invoices optional.

In a lot of context, the reference given by the supplier for the self billed invoice comes from its own system. By default we should not generate this information into Odoo. A flag allows you to adapt this behavior at partner level.